show total count within filter controls component
